requirements.txt
# Telegram Bot Bibliothek
pyTelegramBotAPI

# Webserver für Healthchecks (damit Render nicht einschläft)
flask

# Schnittstelle zu den KI-Modellen
replicate

# Zum Laden von Umgebungsvariablen (.env)
python-dotenv